Changed emulator behavior so that we do not attempt to process aid keys immediately if we are in a non-read state.  Instead, we hang on the the aid and process it when we next enter a read state.
Also fixed bug where we were improperly positioning the cursor at the end of
the write_to_display command.  We were usually positioning to the first
non-bypass field.  This behavior is incorrect when the display is in the
unlocked state.  Instead, we should just leave the cursor where it is.