Include term.h which defines tgetstr
Without this, tgetstr is undefined and by C rules is implicitly defined
to return an int. The compiler then generates code to convert the int
to a pointer and so it throws away the top 32 bits of the pointer that
was actually returned in the register. This causes invalid pointers to
be referenced and segmentation fault immediately.

The same problems happen in tn5250.c because it doesn't know import
cursesterm.h and thus the various functions defined in cursesterm.c are
unknown.