Define AC_CONFIG_AUX_DIR
Without this set, libtoolize will use heuristics to detect it which
looks for install.sh in the current, parent, and grandparent
directories. If any file named install.sh happens to exist in one of
these directories, even if it's not from autotools, it will use that
directory incorrectly, which causes problems. Instead, we set it to the
current directory, where we expect to be.