Fix endian bugs on 64-bit Unix platforms
Unsigned longs are 64-bits on most platforms other than Windows. By
casting to a u_long, we copy 8 bytes over the top of a 4 byte value
causing both a buffer overread and overflow. Besides the obvious
security concerns, it doesn't even work on big endian systems as the
address will get shifted in to the top 4 bytes of the u_long and then
discarded as it gets cast back to an int meaning the address will always
be set to whatever happens to come *after* the actual address.

Similarly, Unix systems define FIONBIO to take an int, while Windows
takes a long. On Windows, a long is always 32-bits, while on Unix it
matches the size of a pointer. On little endian, this is fine because
the low order bits come first so the extra 4 bytes get ignored, but on
big endian it causes the value always to be read as 0. Since ints are
always 32-bit on all platforms we care about (sorry Win16), we can use
that instead.