Refactor OpenSSL autoconf checks
Remove the redundant --with-ssl-dir macro. Instead, the OpenSSL
directory can be specified with --with-ssl. This simplifies handling
since we no longer need to worry about conflicts between these two
settings. Update the doc string to be more descriptive here and follow
autoconf conventions.

Also, instead of doing everything inline in the macro or setting
redundant environment variables, use the predefined autoconf variables
to determine which additional OpenSSL checks we need to run (if any).

Finally, we also need to add the OpenSSL directory found to the library
search path, in case it's in a non-standard path. eg. macOS Homebrew.