Fix assertion in tn5250_session_start_of_header
The if check allows n=0 to continue instead of error early return, but
the assertion is checking that it must be greater than 0. The code
following it handles it perfectly fine, even checking that it is not
zero to prevent an allocation, so I think the assertion is just wrong.