Fix invalid pointer use when setting ip address
If inet_addr returns an IP address of INADDR_NONE in sin_addr.s_addr, the
IP address is looked up using gethostbyname and if that succeeds the
code attempts to copy the IP address returned over to the
sin_addr.s_addr. However, due to a missing ampersand, the s_addr field
will be converted to a pointer and the IP address copied there.

At this point we know s_addr was INADDR_NONE, so the address will be
0xffffffff. This may or may not be a valid/writable address (resulting
in a segmentation fault if not), but either way it won't result in
correct behavior.