Remove Kerberos checks
As far as I can tell, these checks haven't been needed in a long time.
Firstly, Kerberos support was removed from OpenSSL in 1.1.0 in 2016.
Secondly, even on RHEL 7 released in 2014, krb5.h is in /usr/include so
this check is pointless. At this point, all of this is ancient history
so might as well drop. Users can always set CFLAGS to add additional
include paths if they need.