Link with -ltinfo, if needed
Seems at some point libtinfo was split out of libncurses. For
compatibility with existing applications which only expect to link with
libncurses, many distributions replace libncurses.so with a linker
script which adds -ltinfo automatically, but Gentoo does not.

Ideally, we'd use pkg-config here as noted in the Gentoo tracking bug
(https://bugs.gentoo.org/457530), but I'd rather not dive in to that
autoconf mess yet. Instead, we just use libtinfo if it exists.

Fixes #7

Co-authored-by: hijackeel <hijackeel@posteo.net>