Let CMake be able to install things
Currently installs the man pages and command line utilities on Unix; the
Windows GUI version isn't installed. Nor are the data files, Xterm
wrapper script, or the termcap files/key mappings. (Nor is the shared
library, which I don't think has any external consumers anymore, and is
built static in the CMake version.)

Some of these will probably need to be installed again (depending on
what we want to support), and doing so would allow us to get rid of the
autotools stuff.