Fix CI Windows build errors
In d1a47b2c, ws2tcpip.h was pulled in on Windows for getaddrinfo
support. However, this conflicts with winsock.h and we should instead be
using winsock2.h. For some reason, this was not a problem when building
locally, but that could be to differences in Windows SDKs used. On my
Windows 10 machine, I was using 10.0.22621.0 while the GitHub Actions
build was using 10.0.20348.0.

Winsock 2 was added with Windows 95 and at this point there's no reason
to use Winsock 1. However just updating to winsock2.h is not enough,
since windows.h will pull in winsock.h unless WIN32_LEAN_AND_MEAN is
defined (or winsock2.h is included first). Add WIN32_LEAN_AND_MEAN to
pull in less stuff by default, but that means we now need to include
some headers ourselves where previously windows.h did this for us.