Remove xt5250 script
The benefit of this was neglible, and it turns out it's been broken for
a long time anyways. I suspect most invoked the curses frontend
directly, and if we want a better experience, it'd probably be a better
idea to make a dedicated frontend for that than use cruft ol' xterm.